+++ app/controller.py
@@ -1,15 +1,17 @@
 # coding=utf-8
 """Controller File."""
-from zookeeper import init_zk, get_namespace_kafka, get_namespace_saiki, \
-    get_namespace_pemetaan, change_topic_config
-from kazoo.client import NoNodeError, NodeExistsError
+import datetime
 import json
+import logging
+import urllib
+
 from rebalance_partitions import get_zk_dict, generate_json, \
     NotEnoughBrokersException, write_json_to_zk
-import logging
-import datetime
-import urllib
+
 import jmx
+from kazoo.client import NoNodeError, NodeExistsError
+from zookeeper import init_zk, get_namespace_kafka, get_namespace_saiki, \
+    get_namespace_pemetaan
 
 namespace_kafka = get_namespace_kafka()
 namespace_saiki = get_namespace_saiki()
@@ -180,6 +182,41 @@ def update_config(cform):
             cform.segment_jitter_ms.data)
 
     change_topic_config(zk=zk, topic=topic, config_dict=config_dict)
+
+
+def change_topic_config(zk, topic, config_dict):
+    """
+    Update the config for an existing topic and create a change notification
+    so the change will propagate to other brokers
+    """
+    update_topic_config(zk, topic, config_dict)
+    notify_topic_config_change(zk, topic, config_dict)
+    logging.info("created/updated topic config: topic: " + topic +
+                 " , config : " + str(config_dict))
+
+
+def update_topic_config(zk, topic, config_dict):
+    try:
+        zk.create('/config/topics/' + topic,
+                  json.dumps(config_dict, separators=(',', ':'))
+                  .encode('utf-8'), makepath=True)
+    except NodeExistsError:
+        zk.set('/config/topics/' + topic,
+               json.dumps(config_dict, separators=(',', ':')).encode('utf-8'))
+
+
+def notify_topic_config_change(zk, topic, config_dict):
+    node = '/config/changes/config_change_1'
+    content = {'version': 1,
+               'entity_type': 'topics',
+               'entity_name': topic}
+    try:
+        zk.create(node, json.dumps(content, separators=(',', ':'))
+                  .encode('utf-8'), makepath=True)
+    except NodeExistsError:
+        zk.set('/config/changes/config_change_',
+               json.dumps(content, separators=(',', ':'))
+               .encode('utf-8'))
 
 
 def create_topic_entry(topic_name, partition_count, replication_factor):
